Feat/controllers

+++ src/main/java/com/biblioteca/service/BookService.java
@@ -4,10 +4,12 @@ import com.biblioteca.entity.Book;
 import com.biblioteca.entity.BookStatus;
 import com.biblioteca.repository.BookRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import com.biblioteca.exception.ResourceNotFoundException;
 
-import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -25,20 +27,19 @@ public class BookService {
         return bookRepository.save(book);
     }
 
-    @Transactional
+    @Transactional(readOnly = true)
     public Optional<Book> findBookById(Long id) {
         return bookRepository.findById(id);
     }
 
-    @Transactional
-    public List<Book> findAll() {
-        return bookRepository.findAll();
+    @Transactional(readOnly = true)
+    public Page<Book> findAllBooks(Pageable pageable) {
+        return bookRepository.findAll(pageable);
     }
 
     @Transactional
     public Book updateBook(Long id, Book book) {
-        Book existingBook = findBookById(id).orElse(null);
-        if (existingBook != null) {
+        Book existingBook = findBookById(id).orElseThrow(() -> new ResourceNotFoundException("Book not found"));
             existingBook.setTitle(book.getTitle());
             existingBook.setAuthor(book.getAuthor());
             existingBook.setPublicationYear(book.getPublicationYear());
@@ -49,13 +50,14 @@ public class BookService {
             } else {
                 existingBook.setStatus(BookStatus.INDISPONIVEL);
             }
-            return bookRepository.save(book);
-        }
-        return null;
+            return bookRepository.save(existingBook);
     }
 
     @Transactional
     public void deleteBook(Long id) {
+        if (!bookRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Book not found");
+        }
         bookRepository.deleteById(id);
     }
